(feat) move base files and Dolomite to new location
