+++ hummingbot/connector/exchange/dolomite/dolomite_active_order_tracker.pxd
@@ -6,4 +6,3 @@ cdef class DolomiteActiveOrderTracker:
     cdef dict _active_asks
 
     cdef tuple c_convert_snapshot_message_to_np_arrays(self, object message)
-
