+++ hummingbot/connector/exchange/dolomite/dolomite_in_flight_order.pyx
@@ -4,7 +4,7 @@ from decimal import Decimal
 from hummingbot.core.event.events import (OrderFilledEvent, TradeType, OrderType, TradeFee)
 from hummingbot.connector.in_flight_order_base cimport InFlightOrderBase
 from hummingbot.connector.exchange_base cimport ExchangeBase
-from hummingbot.market.dolomite.dolomite_util import (unpad)
+from hummingbot.connector.exchange.dolomite.dolomite_util import (unpad)
 
 s_decimal_0 = Decimal(0)
