+++ hummingbot/connector/exchange/dolomite/dolomite_order_book_tracker_entry.py
@@ -1,6 +1,6 @@
 from hummingbot.core.data_type.order_book import OrderBook
 from hummingbot.core.data_type.order_book_tracker_entry import OrderBookTrackerEntry
-from hummingbot.market.dolomite.dolomite_active_order_tracker import DolomiteActiveOrderTracker
+from hummingbot.connector.exchange.dolomite.dolomite_active_order_tracker import DolomiteActiveOrderTracker
 
 
 class DolomiteOrderBookTrackerEntry(OrderBookTrackerEntry):
